+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## Unreleased
 - Refactoring how the `danger-kotlin-sdk` integrates with the main library
 - Simplified test plugin setup in the project
+- Added test fixtures for `danger-kotlin-sdk`
 
 # 2.0.4
 - Fixed ANR when project is loading in the IDE