+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Refactoring how the `danger-kotlin-sdk` integrates with the main library
 - Simplified test plugin setup in the project
 - Added test fixtures for `danger-kotlin-sdk`
+- Add new `@file:ImportDirectory` kscript annotation for importing entire directories of scripts
+- Disabled all `@Import` and `@ImportDirectory` annotations when loading the script in an IDE context. See https://youtrack.jetbrains.com/issue/KTIJ-16352.
 
 # 2.0.4
 - Fixed ANR when project is loading in the IDE